api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
python-slugify==8.0.4
httpx==0.28.1
loguru==0.7.3
python-dotenv==1.0.1

# dev / test
pytest==8.3.4
pytest-asyncio==0.24.0
